requirement.txt
discord.py
youtube-dl
colorlog
asyncio
cffi
pynacl
aiohttp
python-nmap
random-cat
pythonwhois
pillow
requests